[Xamarin.Android.Build.Tasks] AndroidResource changes do not affect _GenerateJavaStubs
When editing `AndroidResource` files and saving in the IDE, I noticed
that `_GenerateJavaStubs` always runs.

It doesn't seem like we actually need to do that?

The value for `Inputs` makes sense when you look at how the task is
called:

    <Target Name="_GenerateJavaStubs"
        Inputs="$(MSBuildAllProjects);@(_ResolvedUserMonoAndroidAssemblies);$(_AndroidManifestAbs);$(_AndroidBuildPropertiesCache);@(_AndroidResourceDest)">
      <GenerateJavaStubs
        ...
        ResourceDirectory="$(MonoAndroidResDirIntermediate)"
      />
    </Target>

But looking closer, the `ResourceDirectory` is not even used by
`<GenerateJavaStubs/>`. It seems like we can just remove the
`@(_AndroidResourceDest)` item group from the list of `Inputs`.

Testing an `AndroidResource`-change in the Toggl app:

https://github.com/toggl/mobileapp/tree/develop/Toggl.Droid

Before:

    415 ms  GenerateJavaStubs                          1 calls

This task no longer runs. This should be a ~400ms improvement on
design-time builds or incremental builds with `AndroidResource`
changes.

I also added a test checking `AndroidResource` changes and which
targets are skipped during incremental builds. This improvement also
allowed me to decrease the `PerformanceTest` times for
`AndroidResource` changes.