Fix fastdev library side loading
Override directories must be set up both in the debug app helper and the
monodroid runtime as both of them are loaded separately and they do not depend
on each other. This fixes loading of the side loaded shared libraries from the
internal location where they are copied from the override directories.

Add a pipeline step to run tests with fast dev enabled