Fix invalid `mono_aot_mode_name ` in generated symbols
Since https://github.com/xamarin/xamarin-android/commit/decfbccf3e99449c9a8d6afef113c9339f1bc85b, the runtime code seems to expect that `mono_aot_mode_name` is a linked symbol that's written by the build system. It also expects it to be a string that's written, and the code tries to understand in what AOT mode it's supposed to be based on the first character (`n`ormal, `h`ybrid or `f`ull). However, we currently convert the AotMode enum to a `uint` before converting it to a string, meaning instead of writing `normal`, we write `1`. This causes issues with people receiving the following line in their logs :

`W/monodroid(7449): Unknown Mono AOT mode: 1`

This can be confirmed by analyzing the `environment.{env}.s` files generated in the intermediary folder :

```
 .section .rodata..L.str.2,"aMS",@progbits,1
 .type .L.str.2, @object
.L.str.2:
 .asciz "1"
 .size .L.str.2, 2
 .section .data.mono_aot_mode_name,"aw",@progbits
 .global mono_aot_mode_name
mono_aot_mode_name:
 .xword .L.str.2
```

Fix by making sure that we convert the enum to a string instead of the uint. Based on a cursory analysis of the code, this would not affect normal AOT mode, since the call to get_aot_mode seems to only be used for something if aot is not disabled and isn't "normal" (i.e. hybrid or full)