[docs] Add release notes for PR #3899 (#3909)
Context: https://github.com/xamarin/xamarin-android/pull/3899

This is a first example of a new thing I'm trying to help the team add
release notes along the way for all PRs that change user-facing
behavior.  There isn't any terribly special reason I picked this item
as the first example.  It was just an item where I saw a notification
when the PR was merged, so it reminded me that I could add a
release note.

While I was setting this up, I realized that using a single file for
the release notes across all PRs, as suggested in d17e3bd0, would
lead to many unhelpful merge conflicts.

So I think I should revise my suggestion for how to add release notes
going forward.  I think it would be better to use a separate file for
each PR.

This will prevent merge conflicts while also keeping the following
features:

  * PRs can include release notes alongside corresponding code
    changes.
  * Any contributor can propose updates to the release notes for a PR
    after the PR is merged, and the PR author can review the proposed
    updates.
  * Release notes can include images.
  * Release notes can use all of the normal Markdown syntax.

My thought is that each file be named with a "unique" string.  This
could be the PR number, if known, or the branch name used for the PR,
or a GUID.  The filename itself is largely irrelevant; the point is
that it be sufficiently unique so as to avoid potential merge
conflicts with other PRs.