[monodroid] don't start the debugger if it not configured
In a Debug build of `samples/HelloWorld`, startup was taking way longer
than a Release build:

    Debug build:
    ActivityTaskManager: Displayed com.xamarin.android.helloworld/example.MainActivity: +2s689ms
    Release build:
    ActivityTaskManager: Displayed com.xamarin.android.helloworld/example.MainActivity: +678ms

In this case I was launching the app with no debugger attached. Why is
it *so* much slower?

Reading through `debug.cc` and the `start_connection` method, the
return values and comments did not quite match up--or make sense.

It seemed that even though `conn_port` is 0, it was trying to connect
the debugger? I believe this check needed to actually return 2:

    if (!conn_port)
        return 0;

I changed this line, and reworked things to use an `enum` instead of
an `int`. Everything seems clearer now.

We also should check for an empty string from `debug.mono.connect`,
which might actually be the root cause here.

This prevents this log message from being printed at all in this case:

    11-20 16:36:07.392 12306 12306 I monodroid: Connection options: ''

~~ Results ~~

After this change we are now down to:

    ActivityTaskManager: Displayed com.xamarin.android.helloworld/example.MainActivity: +725ms

This is a ~2 second startup improvement for debug builds when no
debugger is attached.