Fix invalid `mono_aot_mode_name ` in generated symbols (#3940)
Since commit decfbccf, the runtime code seems to expect that
`mono_aot_mode_name` is a linked symbol that's generated by the build
system, located in `libxamarin-app.so`.  The runtime code also
expects it to be a string that's written, and the code tries to
understand in what AOT mode it's supposed to be based on the first
character of the string: `n`ormal, `h`ybrid, or `f`ull).

However, we currently convert the `AotMode` enum to a `uint` before
converting it to a string, meaning instead of writing e.g. `normal`,
we instead write `1`.  This causes issues with people receiving the
following line in their logs :

        W/monodroid(7449): Unknown Mono AOT mode: 1

This can be confirmed by analyzing the `environment.{env}.s` files
generated in the intermediary folder :

         .section .rodata..L.str.2,"aMS",@progbits,1
         .type .L.str.2, @object
        .L.str.2:
         .asciz "1"
         .size .L.str.2, 2
         .section .data.mono_aot_mode_name,"aw",@progbits
         .global mono_aot_mode_name
        mono_aot_mode_name:
         .xword .L.str.2

Fix by making sure that we convert the enum to a string instead of a
`uint`.

Based on a cursory analysis of the code, this would not affect normal
AOT mode, since the call to `get_mono_aot_mode()` seems to only be
used for something if aot is not disabled and isn't "normal",
i.e. hybrid or full.