[Xamarin.Android.Build.Tasks] skip _WriteLockFile by default (#3945)
In bf4c3781, we added a warning to detect parallel builds on the same
Xamarin.Android project. This was meant to be a feature for finding
bugs in the IDEs.

However, there appears to be a case this is causing a lot of noise in
Visual Studio for Windows:

1. The user hits F5 / or the play button
2. The IDE runs a build and gets a status back it was successful.
3. The IDE runs the `Install` target.

On step 3, we are randomly getting the warning?

    Xamarin.Android.Common.targets(404,3): warning XA5302: Two processes may be building this project at once. Lock file exists at path: obj\Debug\90\.__lock

It appears that VS is getting a result that the build is finished, but
MSBuild has not cleaned up all `RegisterTaskObject`'s with a `Build`
lifetime:

https://github.com/xamarin/xamarin-android/blob/748dec3f9924836887690655c84472f2890e1f30/src/Xamarin.Android.Build.Tasks/Tasks/WriteLockFile.cs#L55

This might be due to the way VS runs MSBuild, and it runs multiple
MSBuild nodes by default in a solution.

I think we should make a new private `$(_AndroidDetectParallelBuilds)`
MSBuild property that is blank by default. We can turn it on to
diagnose an issue with parallel builds specifically.