[monodroid] check return value of monodroid_get_system_property
In a26e0dbf, I fixed a 2 second delay on startup, which was somewhat
perplexing... It helped by 2 seconds regardless if a debugger was
present or not.

It turned out that a `""` returned from `monodroid_get_system_property`
was the actual root cause of the issue. Luckily, this issue was only
present on d16-5 and master.

We think checking the return value of `monodroid_get_system_property`
is most appropriate. I changed this call to look the same as the rest
of the calls in `monodroid`.