[Xamarin.Android.Build.Tasks] Fix up AndroidStoreUncompressedFileExtensions.
Fixes #3962

Commit d20a991 fixed up an issue where `aapt2` was always compressing
items even if they were defined in `AndroidStoreUncompressedFileExtensions`.
However users are not always including a period in the values they
place in that ItemGroup. In the Linked issue the user was using
`tflite` rather than `.tflite` which `aapt2` is expecrting.

So lets check to see if we have a `.` prefix and if not add one.