[Xamarin.Android.Build.Tasks] Handle Reference Aliases
Fixes #3808

Project Refernces can include [Aliases](https://docs.microsoft.com/en-us/dotnet/api/microsoft.codeanalysis.projectreference.aliases?view=roslyn-dotnet).

```
<ProjectReference Include="..\LibraryA\LibraryA.csproj">
  <Project>{A31DB667-7565-4C13-99CD-82064660960F}</Project>
  <Name>LibraryA</Name>
  <Aliases>LibA</Aliases>
</ProjectReference>
```

When using these users can no longer use the `global` keyword
to reference the types. So

```
	global::LibraryA.Foo.Bar;
```

will become

```
	LibA.LibraryA.Foo.Bar;
```

and an `extern alias LibA;` will need to be added at the very
top of the `.cs` file.
When doing this for a Xamarin.Android project we get the
following build error from the `Resource.designer.cs` file.

```
error CS0400: The type or namespace name 'LibraryA' could not be found in the global namespace (are you missing an assembly reference?)
```

This is because we only use the `global::` when referencing
library types. We do not take into account the `Aliases`
MetaData.

This PR adds support for using the `Aliases` metadata so that
users can now use it in their Xamarin.Android projects.