[Xamarin.Android.Build.Tasks] Fix Library Resource ordering. (#3862)
Fixes: http://work.azdo.io/1009593

Both `aapt` and `aapt2` will use overlay semantics when deciding
which final resource to use.  The order in which the resources are
passed to the tooling is important.

`aapt` uses the *first* item it came across; this is why we always
passed the Application resource directory in first.
For `aapt2` the *last* item is used, and is why we always pass the
Application resource directory last.

However we still use the same ordering for the references as we do
for `aapt`.  As a result if you have duplicate resources in two
library projects, `aapt` and `aapt2` will use different files.

Reverse the order of the libraries we use for `aapt2`.  This ensures
that the behaviour we see in `aapt2` is consistent with `aapt`.

Future work we need to find a way to allow users to change the
ordering of these references.  In Android Studio the order of the
files in the gradle config is used, so users can change the order.
However in MSBuild the order is controlled by MSBuild itself,
and there is no way to override it.