[test] add an assertion checking Debug startup time
In a26e0dbf and 5bfe44e9 we found a performance regression with
startup in Debug builds.

To better catch this in the future, I updated `DebuggingTest`.

If it takes longer than 1 second for the debugger to connect, the test
will fail now.

This worked on my PC with a HAXM emulator, we will see if this number
is too optimistic for our CI servers.