[Xamarin.Android.Build.Tasks] Use Google Manifest Merger. (#3918)
Use the Google Manifest Merger tooling to merge library
`AndroidManifest.xml` files into the App's `AndroidManifest.xml`.
This should allow developers to override and remove some unwanted
permissions which are currently automatically imported from some of
the support libraries.

Documentation on how to use the new manifest merger tooling is at:

	 https://developer.android.com/studio/build/manifest-merge

Because of the limitations in Command line argument lengths under
Windows we have to build our own entry point for the manifest merger
tool.  Our entry point takes a single "response" file as an input,
with each argument needs to be on a separate line for example:

	--main
	AndroidManifest.xml
	--libs
	obj\Debug\lp\0\ji\AndroidManifest.xml:obj\Debug\lp\2\ji\AndroidManifest.xml
	--out
	android\AndroidManifest.xml

We have to put each argument on its own line because of the way the
built in Manifest Merger Argument parser works.