[Xamarin.Android.Build.Tasks] Allow BuildApk to run incrementally.
The `BuildApk` task currently does not work incrementally.
When it is called it throws away the previously built apk
in favour of using the `base` on that `aapt/aapt2` just built.
With incremental builds it is highly likely that only a
few files have actually changed so we are wasting a ton of time.

During my tests on a simple app `BuildApk` generally takes
about 4 seconds to run. This is the same for both clean
and incremental builds.
With these changes, this time could now go down to 600 ms
for an incremental build. Which is a decent improvement.

The new system does not throw away all the work done in
the previous build. Instead it uses the previous output
apk as a base on which to produce the new one.
The output from `aapt/aapt2` is instead `merged` into the
current apk. We do this by checking the `CRC` values of
each entry and only merging in files which have actually
changed. We need to use the `CRC` here because `aapt/aapt2`
is NEVER setting the `ModifiedDate` on the zip items :(.
It seems to always be 01/01/1980.

Once those changes are out of the way , we can now move on
to our custom content. This time we can use the `ModifiedDate`
of both the `ZipEntry` AND the source file. This allows us
to, again, only update the items which have accutally changed.

One final part is we need to keep track of what files are in the
`apk` in the first place and which ones get added to it. This is
what the `existingEntries` List is used for. This allows us to
detect when files are REMOVED from the system. Otherwise we end up
with stale files in the apk from previous builds.

On my tests when changing a single AndroidResource file I got
the following results

	3664 ms BuildApk                  1 calls (Before)
	669 ms BuildApk                  1 calls (After)

Which is a decent improvement.