[Xamarin.Android.Build.Tasks] fix installing the SDK if one doesn't exist (#4003)
Fixes #3996

Commit 008ba98 did not go far enough. Both `ValidateJavaVersion` and `ResolveSdks` also needed the same `ContinueOnError` setting. While the commit did allow users to install components which were missing, it did NOT allow users to install the entire SDK. This is because even when calling's the `InstallAndroidDependencies` target the build would still fail with 

```
error XA5300: The Android SDK Directory could not be found. Please set via /p:AndroidSdkDirectory.
```

What we need to do in this case was include the `ContinueOnError` property on `ValidateJavaVersion` and `ResolveSdks`. This allows the build to continue and call the targets to install the required SDK. This is very handy for CI environments as it means we can install the Android SDK from scratch on a clean machine.