Add unofficial XA project SDK and support for `dotnet build` (#3860)
Context: https://github.com/xamarin/xamarin-android/projects/13
Context: https://devblogs.microsoft.com/dotnet/introducing-net-5/

We are *exploring* what it means for Xamarin.Android to support
[Short-Form projects][0] ("SDK-style" projects) in the context of
.NET 5 and NuGet packages, learning by doing.

To that end, produce a `Xamarin.Android.Sdk*.nupkg` NuGet package,
which is a "NuGet-ized" version of the existing macOS `.pkg` and
Windows `.vsix` installers.  In some ways the package is a standalone
implementation of [MSBuild.Sdk.Extras][1] for Xamarin.Android.

***NOTE***: The NuGet package is ***NOT*** intended for actual use,
and will not be redistributed anytime soon, if ever.  IT WILL NOT BE
SUPPORTED.  This commit only exists so that the Xamarin.Android team
can maintain sanity by *avoiding* a long-lived branch which
inevitably diverges from master.  (Expect changes as we learn more.)

The `$(TargetFramework)` MSBuild property is the only required value:

	<Project Sdk="Xamarin.Android.Sdk/0.0.1">
	  <PropertyGroup>
	    <TargetFramework>MonoAndroid10.0</TargetFramework>
	  </PropertyGroup>
	</Project>

The `Xamarin.Android.Sdk*.nupkg` contents[^0] are a *union* of macOS
and Windows installer contents, which is less than ideal, and will be
addressed later.  Of note:

  * `lib/MonoAndroid` directory, equivalent to the
    `lib/xamarin.android/xbuild-frameworks/MonoAndroid` build tree.

  * `tools` directory, equivalent to the
    `lib/xamarin.android/xbuild/Xamarin/Android` build tree.

  * `Sdk/Sdk.props`, `Sdk/Sdk.targets`: Needed for SDK-style support.
  
  * `targets`: Files referenced by `Sdk/*`.

Such Short-Form projects should be buildable with both `msbuild` and
with `dotnet build`, on both macOS and Windows.

Default Xamarin.Android property values are defined in
`Xamarin.Android.Sdk.props`.  Of note:

  * `$(AndroidApplication)`=True if `Properties\AndroidManifest.xml`
    exists.  (Library projects are assumed to *not* have
    `Properties\AndroidManifest.xml`.)

Default Xamarin.Android item groups are defined in
`Xamarin.Android.Sdk.DefaultItems.props`.  In particular:

  * `@(AndroidResource)`: `Resources\**\*.xml`, `.axml`, `.png`,
    `.jpg`, `.gif`, `.ttf`, `.otf`, `.ttc`, and all `raw\**\*` files.
  * `@(AndroidAsset)`: `Assets\**\*`.

~~ Other changes ~~

The default `$(AndroidSupportedTargetJitAbis)` *xamarin-android*
build property has been updated to include all supported ABIs, as all
of our Azure Pipelines-based builds are already doing this.
This property can still be overridden in `Configuration.props` to
reduce overall build time when desired.

A couple of `StreamWriter` instances have been wrapped in a `using`-
block to fix an issue which would occur only when building with
`dotnet build`.  The `_CompileResources` target was failing during
the `<Aapt2Compile/>` task because we were attempting to use a file
that we were writing to before the underlying `StreamWriter` was
flushed.  For some reason this was not causing an issue when running
on .NET Framework.

The `System.ServiceModel.dll` reference, as well as its single
`PipeException` type usage have been removed for the purposes of
compatibility with `dotnet build`.  The `XA7029` error code was only
added in `d16-4`, and as a result we don't yet have any hits in the
wild.  However, since this was only recently released, I believe it
should be relatively harmless to remove.

We continue to run into test timeouts when running the entire
`Xamarin.Android.Build.Tests` suite against a build tree on Windows.
In some cases our test execution slows down dramatically.  We'll
attempt to address this by reducing the amount of work we do in
parallel.  Our Windows agents seemingly [have 8 cores available][2].
NUnit-Console will attempt to use all cores when running tests in
parallel by default.  Setting the `--workers` argument to 4 should
cut our parallelization in half, and potentially prevent the machine
from running out of resources.

[0]: https://docs.microsoft.com/en-us/visualstudio/msbuild/how-to-use-project-sdk?view=vs-2019
[1]: https://www.nuget.org/packages/MSBuild.Sdk.Extras/
[2]: https://dev.azure.com/devdiv/DevDiv/_settings/agentqueues?agentId=688797&queueId=819&view=capabilities

[^0]: Selected except of `Xamarin.Android.Sdk*.nupkg` contents:

	lib/MonoAndroid/v1.0/Facades/netstandard.dll
	lib/MonoAndroid/v1.0/Facades/…
	lib/MonoAndroid/v1.0/mscorlib.dll
	lib/MonoAndroid/v1.0/…
	lib/MonoAndroid/v1.0/RedistList/FrameworkList.xml
	lib/MonoAndroid/v4.4/Mono.Android.dll
	lib/MonoAndroid/v4.4/OpenTK-1.0.dll
	lib/MonoAndroid/v4.4/RedistList/FrameworkList.xml
	lib/MonoAndroid/v4.4.87/Mono.Android.dll
	lib/MonoAndroid/v4.4.87/RedistList/FrameworkList.xml
	lib/MonoAndroid/v5.0/Mono.Android.dll
	lib/MonoAndroid/v5.0/RedistList/FrameworkList.xml
	lib/MonoAndroid/v5.1/Mono.Android.dll
	lib/MonoAndroid/v5.1/RedistList/FrameworkList.xml
	lib/MonoAndroid/v6.0/Mono.Android.dll
	lib/MonoAndroid/v6.0/RedistList/FrameworkList.xml
	lib/MonoAndroid/v7.0/Mono.Android.dll
	lib/MonoAndroid/v7.0/RedistList/FrameworkList.xml
	lib/MonoAndroid/v7.1/Mono.Android.dll
	lib/MonoAndroid/v7.1/RedistList/FrameworkList.xml
	lib/MonoAndroid/v8.0/Mono.Android.dll
	lib/MonoAndroid/v8.0/RedistList/FrameworkList.xml
	lib/MonoAndroid/v8.1/Mono.Android.dll
	lib/MonoAndroid/v8.1/RedistList/FrameworkList.xml
	lib/MonoAndroid/v9.0/Mono.Android.dll
	lib/MonoAndroid/v9.0/RedistList/FrameworkList.xml
	lib/MonoAndroid/v10.0/Mono.Android.dll
	lib/MonoAndroid/v10.0/RedistList/FrameworkList.xml
	README.md
	Sdk/Sdk.props
	Sdk/Sdk.targets
	targets/Xamarin.Android.Sdk.DefaultItems.props
	targets/Xamarin.Android.Sdk.props
	targets/Xamarin.Android.Sdk.targets
	targets/Xamarin.Android.TargetFrameworkInference.targets
	tools/aapt2.exe
	tools/android-support-multidex.jar
	tools/aprofutil.exe
	tools/aprofutil.pdb
	tools/bcl/Facades/netstandard.dll
	tools/bcl/Facades/…
	tools/bcl/mscorlib.dll
	tools/bcl/…
	tools/Darwin/…
	tools/lib/arm64-v8a/libmono-android.debug.so
	tools/lib/arm64-v8a/…
	tools/lib/armeabi-v7a/…
	tools/lib/x86/…
	tools/lib/x86_64/…
	tools/lib64/libzip.dll
	tools/ndk/…
	tools/proguard/…
	tools/…