[tests] Use non-parallel fixture for InstallAndroidDependencies (#4015)
Context: https://devdiv.visualstudio.com/DevDiv/_build/results?buildId=3307232&view=ms.vss-test-web.test-result-details

We noticed a failure in the `DuplicateReferences` test, which appeared
to be caused by an incorrect Android SDK path being used. MSBuild tests
which build a project will look for the Android SDK in the following
paths:

  1) Value of `$(ANDROID_SDK_PATH)` if set.
  2) The `HKEY_CURRENT_USER\\SOFTWARE\\Novell\\Mono for Android` registry key.
  3) The path we install build tooling into: `%USERPROFILE%\android-toolchain\sdk`.

In the case of the `DuplicateReferences` test failure,
`$(ANDROID_SDK_PATH)` was set to an unexpected value:

  ANDROID_SDK_PATH = F:\A\xamarin-v000006-1\_work\2\s\bin\TestRelease\temp\InstallAndroidDependenciesTest\android-sdk

Presumably this failure occurred because the test ran in parallel with
the `InstallAndroidDependenciesTest` test which modifies the environment.

The fix is to move this test to it's own fixture that will not run in
parallel with other tests.