[Xamarin.Android.Build.Tasks] add feature flag for Google Manifest Merger
Context: https://github.com/xamarin/xamarin-android/pull/3918

In 2c6f5cd1, we added support for the official Google/Android Manifest
Merger. After further QA testing, it seems that a reasonable number of
existing apps require code changes to *appease* the new manifest
merger.

This is a similar situation we were in compared to `aapt2`, where the
suggested changes are good (and usually problems). I think we should
add a feature flag, so we can gradually turn this on. We could add the
feature off by default for one release, and then turn it on by default
for the next.

This PR adds:

* `$(AndroidManifestMerger)` with enum-style values of `legacy` and
  `manifestmerger.jar`.
* Updated a few tests to leverage both code paths.