[Java.Interop-Tests] Use legacy nuget reference
Attempting to use `@(PackageReference)` for `Mono.Linq.Expressions`
continues to fail, even after PR #3995.  Looking at the `msbuild
Xamarin.Android /t:Prepare` output [^0], I see that there is a
`nuget restore Xamarin.Android-Tests.sln`, and it does not always,
and it restores numerous NuGet packages...but not
Mono.Linq.Expressions.

In the interest of any form of progress, use the "legacy"
`packages.config`-style nuget reference, and see if that changes
anything.

[^0]: https://devdiv.visualstudio.com/DevDiv/_build/results?buildId=3318737&view=logs&j=cac0e8d3-0ef5-5d2b-b57e-e8fde7204df3&t=b0ec5c6c-f5a1-5ebf-2068-57a5c78a6a85