[Xamarin.Android.Build.Tasks] add feature flag for Google Manifest Merger (#4022)
Context: https://github.com/xamarin/QualityAssurance/commit/a01b7d9d6a724523878236e49dbb8f69c6dfedc8

In commit 2c6f5cd1, we added support for using the official
Google/Android [Manifest Merger][0].  After further QA testing, it
seems that a reasonable number of existing apps require code changes
to *appease* the new manifest merger:

	error AMM0000: …/QualityAssurance/Samples/android/XAQA.Bug28816/obj/Release/AndroidManifest.xml:4:3-42 Error:
	error AMM0000: 	uses-sdk:minSdkVersion 10 cannot be smaller than version 14 declared in library …/QualityAssurance/Samples/android/XAQA.Bug28816/obj/Release/lp/6/jl/AndroidManifest.xml as the library might be using APIs not available in 10
	error AMM0000: 	Suggestion: use a compatible library with a minSdk of at most 10,
	error AMM0000: 		or increase this project's minSdk version to at least 14,
	error AMM0000: 		or use tools:overrideLibrary="android.support.v4" to force usage (may lead to runtime failures)

This is a similar situation we were in compared to `aapt2`, where the
suggested changes are good (and usually fix problems).  I think we
should add a feature flag, so we can gradually turn this on.
We could add the feature off by default for one release, and then
turn it on by default for the next.

Add a new `$(AndroidManifestMerger)` MSBuild property to control
which `AndroidManifest.xml` merger is used.  The values are:

  * `legacy`: Merge `AndroidManifest.xml` files in the same way we
    did before commit 2c6f5cd1.
  * `manifestmerger.jar`: Use Android's Manifest Merger.

The default value for `$(AndroidManifestMerger)` is `legacy`.

This will change in a future release.

[0]: https://developer.android.com/studio/build/manifest-merge