[ci] Downgrade XA before Windows build (#4060)
Context: https://github.com/xamarin/xamarin-android/pull/4052#issuecomment-567576834

Our `Windows Build and Test` job runs on the same set of VMs that we use
for testing and in some cases this can cause issues. Most recently, I
noticed a failure when attempting to restore NuGet packages for
`Xamarin.Android-Tests.sln`, because the Xamarin.Android version on disk
is used when calculating the NuGet dependecy graph for Android projects.
We should attempt to downgrade XA to the version that shipped with the
instance of VS that we are building with before building on Windows.