[ci] Remove nested xamarin-android submodule (#4091)
Context: https://github.com/xamarin/xamarin-android/pull/4089
Context: ed29bf1d47661e9df8458b5be074ed0369752b0b

In commit ed29bf1d we migrated to a new Azure Pipelines feature that
allows us to clone external sources.  This caused an issue with our
inverted build, as it introduced a nested xamarin-android checkout
which was previously ignored.  When attempting to apply ed29bf1d to
d16-4 -- see also PR #4089 -- our build failed due to an attempt to
restore NuGet packages which are only present (with a proper feed
configuration) in xamarin-android/master:

	Restore failed in 483.85 ms for /Users/runner/runners/2.162.0/work/1/s/xamarin-android/external/monodroid/external/xamarin-android/src/Xamarin.Android.Build.Tasks/Xamarin.Android.Build.Tasks.csproj.
	Committing restore...
	
	Errors in /Users/runner/runners/2.162.0/work/1/s/xamarin-android/external/monodroid/external/xamarin-android/build-tools/Xamarin.Android.Tools.BootstrapTasks/Xamarin.Android.Tools.BootstrapTasks.csproj
	    NU1101: Unable to find package Microsoft.DotNet.GenAPI. No packages exist with this id in source(s): nuget.org
	    NU1101: Unable to find package Microsoft.DotNet.ApiCompat. No packages exist with this id in source(s): nuget.org
	Errors in /Users/runner/runners/2.162.0/work/1/s/xamarin-android/external/monodroid/external/xamarin-android/src/Xamarin.Android.Build.Tasks/Xamarin.Android.Build.Tasks.csproj
	    NU1101: Unable to find package XliffTasks. No packages exist with this id in source(s): nuget.org

Rather than ensuring this submodule is checked out to the right commit,
we'll just nuke the entire nested submodule instead as it is not needed
and should not be used by the inverted build system.