[ci] Report artifact status to the correct repo (#4097)
The `notarize_pkg_upload_storage` job is responsible for posting
artifact metadata to GitHub as a commit status, and the `.vsix`
signing performed by the `queue_vsix_signing` job depends upon the
presence of this GitHub status.

Commit ed29bf1d broke the posting of the status, as it changed the
source checkout behavior for this job.

We report artifact metadata to GitHub with the [Artifacts tool][0],
and this tool must be ran inside the source directory that it is
going to create a GitHub commit status for.  With ed29bf1d, we started
executing this tool within a `release-scripts` checkout rather than a
`xamarin-android` checkout.  This change meant that the required
status was no longer being posted, which in turn caused the
`queue_vsix_signing` job to always fail.

Fix the `queue_vsix_signing` job by cloning both the xamarin-android
and release-scripts repos so that the artifacts tool can run within the
xamarin-android checkout and successfully report a status back to it.

[0]: https://github.com/xamarinhq/build-tasks/tree/master/Artifacts