Clear out the `apiName` for API-29.
As per discussion on 2020-01-08 with @JonDouglas and @kzu (or @joj?), in order to avoid "duplicating" the Android name in the Android SDK drop-down items, the logic will be as follows:

The IDE/etc. will always construct it's own name based on `Android {framework}`.

*If* we specify an `{apiName}`, then `apiName` will be appended within parenthesis.

Thus, API-28 will result in `Android v9.0 (Pie)`.

API-29 will result in `Android v10.0`, *so long as* `apiName` is the empty string.

Clear out `apiname` so that the IDE can start implementing this idea.