Use the compile-time XA SDK version for tests
This generates a .cs file containing an attribute like
the following:

```
[assembly: System.Reflection.AssemblyMetadata("SdkVersion", "10.2.99")]
```

Which can later on be used in the test to initialize the
version to use. If not found, the previous behavior is preserved.

NOTE: if the projects were SDK-style, we could instead use the following
item group instead of a custom target:

```
<ItemGroup Condition="Exists('$(VersionFile)')">
  <AssemblyAttribute Include="System.Reflection.AssemblyMetadata">
    <_Parameter1>SdkVersion</_Parameter1>
    <_Parameter2>$(_SdkVersion)</_Parameter2>
  </AssemblyAttribute>
</ItemGroup>
```