Add unofficial lightweight XA SDK
The existing unofficial XA SDK is a very large package that
might cause issues on low-bandwidth scenarios and can even
hang VS for a long time while it restores it.

As an alternative, this commit provides another SDK, named
`Xamarin.Android.Sdk.Lite`, with the same versioning scheme,
which leverages the existing "fat" SDK targets, but tweaks
them before packaging so that the existing VS-installed
targets are referenced instead. The goal is to basically
leverage all the disk space already taken by the VS
install, instead of downloading everything again. This
means that it also only works for desktop builds using the
desktop `msbuild`, since it relies on the VS installed
paths to locate the Xamarin targets and reference assemblies.